+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## version 0.2.2+ (Unreleased)
 - add support CR, CRLF
 - add some subcommands on v2
+- add --src, --dst options of inshour, insdate subcommand on v2
 
 ## version 0.2.2 (2018-03-07)
 - add `inshour` subcommand on v2